+++ db/migrate/20160622014541_create_reviews.rb
@@ -4,7 +4,7 @@ class CreateReviews < ActiveRecord::Migration
       t.integer :user_id
       t.integer :book_id
       t.text :content
-      t.float :rating
+      t.decimal :rating
 
       t.timestamps null: false
     end
